refactor: extract shared pollForToken utility to eliminate duplication (#25)
* Initial plan

* refactor: extract shared pollForToken utility to eliminate duplication


* chore: remove accidental package-lock.json and add to .gitignore


---------
